Uma Equipe Pode Ter Varios gestores, um vendedor so pode pertencer a uma equipe

+++ backend/src/main/java/com/devsuperior/ControleDeVendas/services/SaleService.java
@@ -41,6 +41,10 @@ public class SaleService {
 			Page<Sale> page = repository.findBySeller(user, pageable);
 			return page.map(x -> new SaleDTO(x));
 		}
+		else if (user.hasRole("ROLE_MANAGER")) {
+			Page<Sale> page = repository.findByManage(user.getId(), pageable);
+			return page.map(x -> new SaleDTO(x));
+		}
 		else {
 			Page<Sale> page = repository.findAll(pageable); 
 			return page.map(x -> new SaleDTO(x));
@@ -70,21 +74,8 @@ public class SaleService {
 	@Transactional
 	public SaleDTO insert(SaleDTO dto) {
 		User user = authService.authenticated();
-		Sale entity = new Sale();
-		entity.setDate(dto.getDate());
-		entity.setVisited(dto.getVisited());
-		entity.setDeals(dto.getDeals());
-		entity.setAmount(dto.getAmount());
-		entity.setSeller(userRepository.getOne(user.getId()));
-		entity = repository.save(entity);
-		return new SaleDTO(entity);
-	}
-	
-	@Transactional
-	public SaleDTO update(Long id, SaleDTO dto) {
-		User user = authService.authenticated();
-		try {
-			Sale entity = repository.getOne(id);
+		if (user.hasRole("ROLE_SELLER")) {
+			Sale entity = new Sale();
 			entity.setDate(dto.getDate());
 			entity.setVisited(dto.getVisited());
 			entity.setDeals(dto.getDeals());
@@ -93,21 +84,53 @@ public class SaleService {
 			entity = repository.save(entity);
 			return new SaleDTO(entity);
 		}
-		catch (EntityNotFoundException e) {
-			throw new ResourceNotFoundException("Entity not Found " + id);
+		else {
+			throw new UnauthorizedException("Unauthorized User");
 		}
+		
+	}
+	
+	@Transactional
+	public SaleDTO update(Long id, SaleDTO dto) {
+		User user = authService.authenticated();
+		if (user.hasRole("ROLE_SELLER")) {
+			try {
+				Sale entity = repository.getOne(id);
+				entity.setDate(dto.getDate());
+				entity.setVisited(dto.getVisited());
+				entity.setDeals(dto.getDeals());
+				entity.setAmount(dto.getAmount());
+				entity.setSeller(userRepository.getOne(user.getId()));
+				entity = repository.save(entity);
+				return new SaleDTO(entity);
+			}
+			catch (EntityNotFoundException e) {
+				throw new ResourceNotFoundException("Entity not Found " + id);
+			}
+		}
+		else {
+			throw new UnauthorizedException("Unauthorized User");
+		}
+		
 	}
 	
 	public void delete(Long id) {
-		try {
-			repository.deleteById(id);
+		User user = authService.authenticated();
+		if (user.hasRole("ROLE_SELLER")) {
+			try {
+				repository.deleteById(id);
+			}
+			catch(EntityNotFoundException e) {
+				throw new ResourceNotFoundException("Entity not Found " + id);
+			}
+			catch(DataIntegrityViolationException e) {
+				throw new DatabaseException("Integrity violation");
+			}
 		}
-		catch(EntityNotFoundException e) {
-			throw new ResourceNotFoundException("Entity not Found " + id);
+		else {
+			throw new UnauthorizedException("Unauthorized User");
 		}
-		catch(DataIntegrityViolationException e) {
-			throw new DatabaseException("Integrity violation");
-		}
+		
 	}
 
 }
